create a separate module for example creation
